refactor(arch): move http/ to transport/http/ and split handlers/validators
Move src/server/http/* to src/transport/http/ and split:
- handlers.ts → handlers/{sessions,permissions,events,settings,system,projects}.ts
- validators.ts → validators/{common,sessions,settings}.ts
- auth.ts, cors.ts, json.ts → middlewares/
- server.test.ts → __tests__/server.test.ts (integration)

Co-located unit tests follow their source. Splits clarify domain
boundaries even where files were under the AGENTS.md §3 ~300 LoC
threshold.

The __dirname arithmetic in handlers/system.ts uses a transient path
(../../../server/dashboard) that targets the dashboard at its current
location. Commit 5 moves the dashboard out and updates the path to
../../../dashboard.

Behavior preserved: 377 tests pass.

Part 4 of 6 (see docs/REFACTOR-2026-04-architecture.md).
