+++ src/server/index.ts
@@ -13,7 +13,7 @@ import { writeState, clearState, globalStatePath } from "../core/state/store"
 import { existsSync } from "fs"
 import { writeBanner } from "../infra/banner/writer"
 import { createNotificationService } from "../notifications/pipeline"
-import { createRemoteServer } from "./http/server"
+import { createRemoteServer } from "../transport/http/server"
 import { opencodeIntegration } from "../integrations/opencode/index"
 import { codexIntegration } from "../integrations/codex/index"
 import { createLogger } from "../infra/logger/index"
